Parsing for Stories PS2, PSP
Welp, after working back and forth between Stories/Grand Theft Auto IV, and re-ing .MDL and .WDR formats(with more luck with the latter at first) for the first half of may, and then working on models/maps for the second half, I managed to have a Eureka moment - so I spent all June working on MDLs. There's a lot to say that I can't all put here. Firstly, Stories/Leeds documentation is scarce, more so than RW or RAGE. Secondly - as well as last of all for now - it's amazing that it's been nearly 20 years for both GTA Stories games, and not a single crumb of an IO or decent documentation for the PSP formats... well, that ends now.