Import mesh(es) for Stories PSP
So, I know I got a handful of resources in the scripts; articles relating to Stories, Manhunt2, or just Leeds in general, but the most helpful for PSP was definitely The_Hero's work on re'd RW/Leeds Engine, which is a converter, rather than an importer with GUI. Still helpful though. There's still lots of things to do however; materials need to be assigned to mesh based on their ID, the armature needs to be fixed, and some tristrip split sections(specifically the last one in a mesh) get skipped.

So far so good, though. Pretty good for a month and a half of work so far.