CW: Read NumLights, create UV Map, assign materials
Forgot to mention: read lights based on value of NumLights. created UV map so we can properly assign materials & images, but seem to be having problems with the uvs.
An example is worldblock65 has the same texcoords as it does in DK22Pacs CW Mobile Explorer, yet the textures look completely scrambled.  Should probably fix asap